Remove bs shader code from material import. Pull out properties from mat loop.
