+++ io_scene_nif/modules/property/shader/shader_import.py
@@ -36,10 +36,20 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 # ***** END LICENSE BLOCK *****
+import bpy
 from pyffi.formats.nif import NifFormat
+
+from io_scene_nif.modules.object.object_import import Object
+from io_scene_nif.modules.property.texture.texture_import import Texture
+from io_scene_nif.utility import nif_utils
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class BSShader:
+
+    def __init__(self):
+        self.dict_materials = {}
+        self.texturehelper = Texture()
 
     @staticmethod
     def import_shader_types(b_obj, b_prop):
@@ -73,3 +83,100 @@ class BSShader:
             sf_index = b_prop.shader_flags_2._names.index(b_flag_name_2)
             if b_prop.shader_flags_2._items[sf_index]._value == 1:
                 b_obj.niftools_shader[b_flag_name_2] = True
+
+    def import_bsshader_material(self, bs_shader_property, bs_effect_shader_property, n_alpha_prop):
+        material_hash = self.get_bsshader_hash(bs_shader_property, bs_effect_shader_property)
+        try:
+            return self.dict_materials[material_hash]
+        except KeyError:
+            pass
+
+        # name unique material
+        name = Object.import_name(bs_shader_property)
+        if not name:
+            name = (self.nif_import.active_obj_name + "_nt_mat")
+        b_mat = bpy.data.materials.new(name)
+
+        if bs_shader_property:
+            self.texturehelper.import_bsshaderproperty_textures(b_mat, bs_shader_property)
+        if bs_effect_shader_property:
+            self.texturehelper.import_bseffectshaderproperty_textures(b_mat, bs_effect_shader_property)
+
+        # shader based properties
+        if bs_shader_property:
+
+            # Diffuse color
+            if bs_shader_property.skin_tint_color:
+                b_mat.diffuse_color.r = bs_shader_property.skin_tint_color.r
+                b_mat.diffuse_color.g = bs_shader_property.skin_tint_color.g
+                b_mat.diffuse_color.b = bs_shader_property.skin_tint_color.b
+                b_mat.diffuse_intensity = 1.0
+
+            if (b_mat.diffuse_color.r + b_mat.diffuse_color.g + b_mat.diffuse_color.g) == 0:
+                b_mat.diffuse_color.r = bs_shader_property.hair_tint_color.r
+                b_mat.diffuse_color.g = bs_shader_property.hair_tint_color.g
+                b_mat.diffuse_color.b = bs_shader_property.hair_tint_color.b
+                b_mat.diffuse_intensity = 1.0
+
+            # Emissive
+            b_mat.niftools.emissive_color.r = bs_shader_property.emissive_color.r
+            b_mat.niftools.emissive_color.g = bs_shader_property.emissive_color.g
+            b_mat.niftools.emissive_color.b = bs_shader_property.emissive_color.b
+            b_mat.emit = bs_shader_property.emissive_multiple
+
+            # Alpha
+            if n_alpha_prop:
+                b_mat = self.set_alpha_bsshader(b_mat, bs_shader_property, n_alpha_prop)
+
+            # gloss
+            b_mat.specular_hardness = bs_shader_property.glossiness
+
+            # Specular color
+            b_mat.specular_color.r = bs_shader_property.specular_color.r
+            b_mat.specular_color.g = bs_shader_property.specular_color.g
+            b_mat.specular_color.b = bs_shader_property.specular_color.b
+            b_mat.specular_intensity = bs_shader_property.specular_strength
+
+            # lighting effect
+            b_mat.niftools.lightingeffect1 = bs_shader_property.lighting_effect_1
+            b_mat.niftools.lightingeffect2 = bs_shader_property.lighting_effect_2
+
+        if bs_effect_shader_property:
+            # Alpha
+            if n_alpha_prop:
+                b_mat = self.set_alpha_bsshader(b_mat, bs_shader_property, n_alpha_prop)
+
+            if bs_effect_shader_property.emissive_color:
+                b_mat.niftools.emissive_color.r = bs_effect_shader_property.emissive_color.r
+                b_mat.niftools.emissive_color.g = bs_effect_shader_property.emissive_color.g
+                b_mat.niftools.emissive_color.b = bs_effect_shader_property.emissive_color.b
+                b_mat.niftools.emissive_alpha = bs_effect_shader_property.emissive_color.a
+                b_mat.emit = bs_effect_shader_property.emissive_multiple
+            b_mat.niftools_alpha.textureflag = bs_effect_shader_property.controller.flags
+
+    def set_alpha_bsshader(self, b_mat, shader_property):
+        NifLog.debug("Alpha prop detected")
+        b_mat.use_transparency = True
+        b_mat.alpha = (1 - shader_property.alpha)
+        b_mat.transparency_method = 'Z_TRANSPARENCY'  # enable z-buffered transparency
+        return b_mat
+
+    def get_bsshader_hash(self, bs_shader_property, bs_effect_shader_property):
+        return (bs_shader_property.get_hash()[1:] if bs_shader_property else None,  # skip first element, which is name
+                bs_effect_shader_property.get_hash() if bs_effect_shader_property else None)
+
+    # TODO [shader] Move move out when nolonger required to reference
+    def find_bsshaderproperty(self, n_block):
+        # bethesda shader
+        bsshaderproperty = nif_utils.find_property(n_block, NifFormat.BSShaderPPLightingProperty)
+        if not bsshaderproperty:
+            bsshaderproperty = nif_utils.find_property(n_block, NifFormat.BSLightingShaderProperty)
+
+        if bsshaderproperty:
+            for textureslot in bsshaderproperty.texture_set.textures:
+                if textureslot:
+                    self.bsShaderProperty1st = bsshaderproperty
+                    break
+            else:
+                bsshaderproperty = self.bsShaderProperty1st
+        return bsshaderproperty
